chore: move out name space, key, value
